перемещение дз в папки
