--- task 2/Descriprion.txt
@@ -1,2 +1,0 @@
-2) Скважина фонтанирует и давление несколько избыточно, что в итоге приводит к тому, что режим нестабилен и изменится до стабильного (пересечение IPR и VLP кривой)
-3) Скважина работает в режиме близком к эффективному, о чём говорит близость к пересечению кривых
