--- task 2/result2.json
@@ -1,33 +1,0 @@
-{
-      "1 task": {
-			"gas_fraction_standing": 39.36, 
-			"q_mix_standing": 86.95587224, 
-			"gas_viscosity_standing": 0.018444363,
-			"gas_density_standing": 156.6409325,
-			"water_viscosity_standing": 1.137528857,
-			"water_density_standing": 1020.19896,
-			"oil_viscosity_standing": 2.660557576,
-			"oil_density_standing": 667.324407,
-			"mixture_viscosity_standing":1.137528857,
-			"mixture_density_standing": 749.7021954,
-			"gas_fraction_mccain": 33.06,
-			"q_mix_mccain": 85.26431038,
-			"gas_viscosity_mccain": 0.0184,
-			"gas_density_mccain": 156.64,
-			"water_viscosity_mccain": 1.1375:,
-			"water_density_mccain": 1020.19896,
-			"oil_viscosity_mccain": 1.418713909,
-			"oil_density_mccain": 630.3757542,
-			"mixture_viscosity_mccain": 1.137528857,
-			"mixture_density_mccain": 775.5918876},
-	  },
-      "2 task": {
-            "p_wf": 212.97
-			"q_liq_vlp": "[0, 50, 100, 150, 200, 250, 300, 350, 400, 450, 500]",
-            "P_wf_vlp": "[271.0378837, 228.8748522, 251.621691, 275.7724843, 298.8288218, 320.2333253, 340.1989228, 358.9882097, 376.861473, 711.539173, 712.8502632]"
-      },
-      "3 task": {
-            "h_m": "[0, 198.6547368, 397.3094737, 595.9642105, 794.6189474, 993.2736842, 1191.928421, 1390.583158, 1589.237895, 1787.892632, 1986.547368, 2185.202105, 2383.856842, 2582.511579, 2781.166316, 2979.821053, 3178.475789, 3377.130526, 3575.785263, 3774.44]",
-            "P_wf": "[4, 4, 4, 4, 4, 4, 4, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5]"
-      }
-}
