--- task 3/result3.json
@@ -1,5 +1,0 @@
-{
-	"ex1":{"p_crit":8.32, "q_crit":175.1},
-	"ex2":{"d_ch":16.35}
-
-}
