+++ task 3/Голубев/result3.json
@@ -1,0 +1,5 @@
+{
+	"ex1":{"p_crit":8.32, "q_crit":175.1},
+	"ex2":{"d_ch":16.35}
+
+}
