+++ task 3/Голубев/техрежим_газлифт (1).xlsx
